pyproject.toml
[tool.poetry]
name = "playground-backend-github"
version = "0.1.0"
description = ""
authors = ["Leon <i@leon.io>"]

[tool.poetry.dependencies]
python = "^3.9"
sanic = "^21.12.1"
tortoise-orm = "^0.18.1"
sanic-jwt = "^1.7.0"
aiomysql = "^0.0.22"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
